requirements.txt
aiofiles==22.1.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
choicelib==0.1.5
exceptiongroup==1.1.3
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
multidict==6.0.4
numpy==1.25.2
pydantic==1.10.12
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
transliterate==1.10.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vbml==1.1.post1
vk-api==11.9.9
vkbottle==4.3.12
vkbottle-types==5.131.146.16
watchfiles==0.18.1
yarl==1.9.2
